# Conflicts:
#	src/main/java/hello/fclover/config/SecurityConfig.java
#	src/main/resources/static/images/user/favicon.png
#	src/main/resources/static/images/user/redHeart.png
#	src/main/resources/static/sql/address_book.sql
#	src/main/resources/static/sql/cart.sql
#	src/main/resources/static/sql/goods.sql
#	src/main/resources/static/sql/wish.sql

+++ src/main/resources/static/sql/member.sql
@@ -1,7 +1,6 @@
 CREATE TABLE member
 (
-
-    member_no       INT AUTO_INCREMENT PRIMARY KEY,
+    member_no       BIGINT AUTO_INCREMENT PRIMARY KEY,
     member_id       VARCHAR(100) NOT NULL UNIQUE,
     password        VARCHAR(100),
     name            VARCHAR(100) NOT NULL,
